requirements.txt
gboml==0.1.3
gurobipy==11.0.0
matplotlib-inline==0.1.6
